refactor PeerConnections to api
